requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0
matplotlib>=3.5.0
einops>=0.6.0
tqdm>=4.64.0
scikit-learn
scipy